Got it working! :)
Updates:
main: I removed or commented out unnecessary "food" items and updated variable name to match throughout the program.

cycle2: Updated the "Cycle2" class to call for the Actor class; this is a more appropriate as it allows for it to inherit more attributes and behaviors. I added in missing functions from the "Cycle" class because this is partially what kept the second snake from moving properly.

score: added text for Player Two's score. This still needs to be updated to keep track of scores.

control_actors_action: added in code to establish two separate directions. This allows both snakes to be controlled by their respective keys.

draw_actors_action: updated code to allow for population of both scores, snakes, and segments.

handle_collisions_action: removed unnecessary code for food collision. Added code for head and segment collision for the snakes.